Refactor and commonize compiler.py, create symbols.py
